jQuery Scala Stack - Moved packages - Result type - Refactoring logic
