+++ src/main/scala/datastore/entity/ResourceEntity.scala
@@ -1,4 +1,4 @@
-package datastore
+package datastore.entity
 
 import com.googlecode.objectify.annotation.Entity
 import com.googlecode.objectify.annotation.Cache
