+++ src/main/webapp/scalajs/script/FrontendLogic.scala
@@ -1,18 +1,23 @@
-package webapp.script
+package scalajs
 
 import shared._
 
 trait FrontendLogic {
   
-  def create(getPostValue : () => Integer, postAction : Resource => Unit) = {
-    val number = getPostValue()
+  def create(getPostValue : () => Integer, postAction : Resource => Unit) : Result[Throwable, Long] = {
     
-    val test = new Resource(x = number)
-    
-    // Validated with shared logic
-    ResourceValidator.apply(test)
-    
-    postAction(test)
-    
+    try {
+      val number = getPostValue()
+      
+      val test = new Resource(x = number)
+      
+      // Validated with shared logic
+      ResourceValidator.apply(test)
+      
+      postAction(test)
+    } catch {
+      case t : Throwable => 
+        Failure(t)
+    }
   }
 }
